backend/requirements.txt
fastapi>=0.115.12
uvicorn[standard]>=0.34.2
sqlalchemy>=2.0.41
psycopg2-binary>=2.9.10
python-dotenv>=1.1.0
pydantic>=2.0

langchain>=0.3.25
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.7.0